# cms-assignment-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS Assignment API",
        "description": "CMS Backend API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local API"
        }
    ],
    "paths": {
...
        "/pages": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "List Pages",
                "description": "GET /pages",
                "operationId": "be285396fe3a2c5171e59e42660abcc8",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search pages by title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "menu_id",
                        "in": "query",
                        "description": "Filter by menu id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by page status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "published"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pages retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Create Page",
                "description": "POST /pages",
                "operationId": "806faa5fd5ff69e010b683fe7fe07592",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "menu_id",
                                    "title",
                                    "body",
                                    "status"
                                ],
                                "properties": {
                                    "menu_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "About Us"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "about-us"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "This is page content."
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "published"
                                    },
                                    "publish_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Page created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pages/{page}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Show Page",
                "description": "GET /pages/{page}",
                "operationId": "35b609a40fc6d342e0f3a53d33cc3f56",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page retrieved successfully"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pages"
                ],
                "summary": "Update Page",
                "description": "PUT /pages/{page}",
                "operationId": "9039f2e5643a1df726ef741f246726f9",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "menu_id",
                                    "title",
                                    "body",
                                    "status"
                                ],
                                "properties": {
                                    "menu_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Page"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "updated-page"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Updated page content."
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "draft"
                                    },
                                    "publish_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Page updated successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pages"
                ],
                "summary": "Delete Page",
                "description": "DELETE /pages/{page}",
                "operationId": "57076e2c7a0ba1380ff9a932958859d1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page deleted successfully"
                    },
                    "404": {
                        "description": "Page not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}